Changed the integer datatypes in the methods of the GameObject class to uint32_t and int32_t

+++ include/GameObject.h
@@ -5,9 +5,9 @@
 class GameObject {  
     
     public:
-        void Scale(int8_t scalar);
-        void Translate(int8_t delta_x, int8_t delta_y);
-        void PlaceAt(uint16_t _x, uint16_t _y);
+        void Scale(int32_t scalar);
+        void Translate(int32_t delta_x, int32_t delta_y);
+        void PlaceAt(uint32_t _x, uint32_t _y);
         
         SDL_Rect* GetRect();
         SDL_Texture** GetTexture();
