+++ include/Text.h
@@ -18,9 +18,9 @@ public:
     Text(std::string _text, uint8_t size, TTF_Font*& _font, SDL_Color& _color, SDL_Renderer*& renderer);
     ~Text();
     
-    void Scale(int8_t scalar);
-    void Translate(int8_t delta_x, int8_t delta_y);
-    void PlaceAt(uint16_t _x, uint16_t _y);
+    void Scale(int32_t scalar);
+    void Translate(int32_t delta_x, int32_t delta_y);
+    void PlaceAt(uint32_t _x, uint32_t _y);
 
     SDL_Rect* GetRect();
     SDL_Texture** GetTexture();
